requirements.txt
fastapi
uvicorn[standard]
httpx
playwright
openai
pydantic
python-dotenv
pdfplumber
pytest
pytest-asyncio
gunicorn
